requirements-ms.txt
c8y-api
flask
apscheduler
python-dotenv
